44 inheritance  developing a second subclass
