--- type_script_crash_course/05-inheritance/01-overview-with-sape/.gitignore
@@ -1,1 +1,0 @@
-dist/*.*
