+++ type_script_crash_course/05-inheritance/03-developing-a-subclass/src/Circle.ts
@@ -6,8 +6,16 @@ export class Circle extends Shape {
         super(_theX, _theY);
     }
 
+    public get radius(): number {
+        return this._radius;
+    }
+
+    public set radius(value: number) {
+        this._radius = value;
+    }
+
     public getInfo(): string {
-        return super.getInfo() + `, radius=${this._radius}`;
+        return super.getInfo() + `, radius=${this.radius}`;
     }
