+++ type_script_crash_course/05-inheritance/03-developing-a-subclass/tsconfig.json
